Add error message if hotkey couldn't be read
I'm not sure what causes this, but I had a couple of workflows that
didn't use any of the recognized 'hotmod' values. They were fairly old
workflows, so perhaps the hotmod values have changed since then. At any
rate, this protects against that and shows an error message in the
markdown output.